refactor: abstract http engine tracing
